GML: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31608 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gml/gmlfeatureclass.cpp
@@ -60,7 +60,7 @@ GMLFeatureClass::GMLFeatureClass( const char *pszName ) :
     m_bSRSNameConsistent(true)
 {
     m_pszName = CPLStrdup( pszName );
-    n_nNameLen = strlen( m_pszName );
+    n_nNameLen = static_cast<int>(strlen( m_pszName ));
 }
 
 /************************************************************************/
@@ -267,7 +267,7 @@ void GMLFeatureClass::SetElementName( const char *pszElementName )
 {
     CPLFree( m_pszElementName );
     m_pszElementName = CPLStrdup( pszElementName );
-    n_nElementNameLen = strlen(pszElementName);
+    n_nElementNameLen = static_cast<int>(strlen(pszElementName));
 }
 
 /************************************************************************/
@@ -447,7 +447,7 @@ bool GMLFeatureClass::InitializeFromXML( CPLXMLNode *psRoot )
 /* -------------------------------------------------------------------- */
     CPLFree( m_pszName );
     m_pszName = CPLStrdup( CPLGetXMLValue( psRoot, "Name", NULL ) );
-    n_nNameLen = strlen(m_pszName);
+    n_nNameLen = static_cast<int>(strlen(m_pszName));
 
     SetElementName( CPLGetXMLValue( psRoot, "ElementPath", m_pszName ) );
